extracting out parallelize_dataframe, parallizing synonyms
