+++ pipeline/data_merging/brca_pseudonym_generator.py
@@ -3,7 +3,6 @@ import os
 import pickle
 import subprocess
 import tempfile
-from multiprocessing.pool import ThreadPool
 
 import click
 import hgvs.assemblymapper
@@ -12,13 +11,14 @@ import hgvs.normalizer
 import hgvs.parser
 import hgvs.projector
 import hgvs.validator
-import numpy as np
+
 import pandas as pd
 from hgvs.exceptions import HGVSError
 
 from common import config
 from common.hgvs_utils import HgvsWrapper
 from common.variant_utils import VCFVariant
+from common import utils
 
 SYNONYMS_FIELD = 'Synonyms'
 
@@ -27,16 +27,6 @@ VAR_OBJ_FIELD = 'var_objs'
 NEW_SYNONYMS_FIELD = 'new_syns'
 TMP_CDNA_UNORM_FIELD = 'tmp_hgvs_cdna_unorm'
 TMP_CDNA_NORM_FIELD = 'tmp_hgvs_cdna_norm'
-
-# TODO: use! (check pathos)
-def parallelize_dataframe(df, func, n_cores=4):
-    df_split = np.array_split(df, n_cores)
-
-    pool = ThreadPool(n_cores)
-    df = pd.concat(pool.map(func, df_split))
-    pool.close()
-    pool.join()
-    return df
 
 
 def _get_cdna(df, pkl, hgvs_proc):
@@ -66,7 +56,7 @@ def _get_cdna(df, pkl, hgvs_proc):
         pickle_dict = pickle.load(open(pkl, 'rb'))
         s_cdna = pd.Series([ pickle_dict[str(v)]  for v in var_objs  ])
     else:
-        s_cdna = parallelize_dataframe(df, lambda dfx: dfx.apply(from_field_or_compute, axis=1), 4)
+        s_cdna = utils.parallelize_dataframe(df, lambda dfx: dfx.apply(from_field_or_compute, axis=1), 4)
 
         if pkl:
             pickle_dict = {str(v): c for (c, v) in zip(s_cdna, var_objs)}
@@ -171,8 +161,6 @@ def main(input, output, pkl, log_path, config_file, resources):
 
     df = pd.read_csv(input, sep='\t')
 
-    df = df.iloc[0:1000] # TODO: remove
-
     df[VAR_OBJ_FIELD] = df.apply(lambda x: VCFVariant(x['Chr'], x['Pos'], x['Ref'], x['Alt']), axis=1)
 
     # CDNA conversions
@@ -201,11 +189,10 @@ def main(input, output, pkl, log_path, config_file, resources):
     df['pyhgvs_Protein'] = (df[TMP_CDNA_NORM_FIELD].
                             apply(lambda hgvs_cdna: hgvs_proc._to_protein(hgvs_cdna)))
 
-    df[NEW_SYNONYMS_FIELD] = df.apply(lambda s: get_synonyms(s, hgvs_proc, syn_ac_dict), axis=1)
+    df[NEW_SYNONYMS_FIELD] = utils.parallelize_dataframe(df, lambda d: d.apply(lambda s: get_synonyms(s, hgvs_proc, syn_ac_dict), axis=1), 4)
 
     df[SYNONYMS_FIELD] = df[SYNONYMS_FIELD].fillna('').str.strip()
 
-    # TODO: rename to generated_syonynms?
     # merge existing synonyms with generated ones and sort them
     df[SYNONYMS_FIELD] = df.apply(_merge_synonyms, axis=1)
